Add null mending dirs in report and fix report parameters when the first iteration fails in diag-exporter

+++ cmender/src/main/java/pt/up/fe/specs/cmender/Launcher.java
@@ -55,6 +55,9 @@ public class Launcher {
 
         // delete the mending directory
         for (var mendingDirData : mendingDirDatas) {
+            if (mendingDirData == null) {
+                continue;
+            }
             try {
                 FileUtils.deleteDirectory(new File(mendingDirData.dirPath()));
             } catch (IOException e) {
